🦄 [refactor]: fetch the deal of the month from db; improved ui [Refactor]: Deal of the month #37
