+++ component/shared/promo/index.ts
@@ -1,1 +1,2 @@
-export { default as DealCountdown } from './deal-countdown'
+export { default as DealCountdown } from './promo-dotm-countdown'
+export { default as PromoSwiper } from './promo-swiper'
